+++ Assets/Scripts/Gates/GateOperationType.cs
@@ -1,0 +1,9 @@
+namespace InfiniStacker.Gates
+{
+    public enum GateOperationType
+    {
+        Add,
+        Subtract,
+        Multiply
+    }
+}